requirements.txt
jupyter~=1.1.1
tqdm~=4.66.5

pillow~=11.0.0
opencv-python~=4.10.0.84

numpy~=1.26.4

torch~=2.5.0
torchvision~=0.20.0
lightning~=2.4.0
segmentation-models-pytorch~=0.3.4
ultralytics~=8.3.25

wandb~=0.18.5
torchview~=0.2.6
graphviz~=0.20.3
matplotlib~=3.9.2

goopylib~=2.1.0
